データ共有

+++ Src/Audio/Audio.cpp
@@ -138,7 +138,7 @@ bool LoadWaveFile(HANDLE hFile, WF& wf, std::vector<UINT32>& seekTable, std::vec
     }
 
     if (chunk.tag == FOURCC_FORMAT_TAG) {
-      if (!Read(hFile, &wf.u, std::min(chunk.size, sizeof(WF::U)))) {
+      if (!Read(hFile, &wf.u, std::min(chunk.size, static_cast<uint32_t>(sizeof(WF::U))))) {
         break;
       }
       switch (GetWaveFormatTag(wf.u.ext)) {
